pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "protein_annotator"
authors = [
    { name = "Lucas Sandoval", email = "x@x.com" },
    { name = "Mauro Portillo", email = "x@x.com" },
    { name = "Nicolas Raño", email = "x@x.com" },
    { name = "Pablo Spizzamiglio", email = "pablo.spizzamiglio@unq.edu.ar" },
]
description = "Protein Annotator"
requires-python = ">=3.8"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
license = { file = "LICENSE" }
readme = "README.md"
dependencies = [
    "biopython >= 1.81",
    "httpx >= 0.24.0",
    "pydantic >= 2.0.0",
    "typing-extensions >= 4.8.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "black ~= 23.11.0",
    "isort ~= 5.12.0",
    "mypy ~= 1.7.0",
    "flake8",
    "flake8-bugbear",
]
docs = [
    "mkdocs",
    "mkdocs-embed-external-markdown",
    "mkdocs-exclude",
    "mkdocs-material",
    "mkdocs-redirects",
    "mkdocs-simple-hooks",
    "mkdocstrings-python",
    "tomli",
]
release = ["build", "wheel", "twine"]
test = ["coverage[toml]", "pytest", "pytest-mock", "pytest-pretty"]

[tool.setuptools.package-data]
protein_annotator = ["py.typed"]

[tool.setuptools_scm]

[tool.black]
line-length = 88
skip-string-normalization = false
target_version = ["py38", "py310", "py311", "py312"]

[tool.isort]
line_length = 88
multi_line_output = 3
combine_as_imports = true
profile = "black"
extra_standard_library = ["typing_extensions"]

[tool.mypy]
python_version = "3.8"
platform = "linux"
pretty = true
strict = true
show_column_numbers = true
show_error_codes = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = ["-vvv", "-s"]
testpaths = ["tests"]

[tool.coverage.paths]
protein_annotator = ["src/protein_annotator"]
tests = ["tests"]

[tool.coverage.run]
command_line = "-m pytest"
context = "protein_annotator"
branch = true
source_pkgs = ["protein_annotator"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    '\#\s*pragma: no cover',
    '^\s*raise NotImplementedError\b',
    "if typing.TYPE_CHECKING:",
]

[tool.coverage.html]
show_contexts = true
